+++ contracts/logic/StakingPoolLogicV2.sol
@@ -75,14 +75,14 @@ library StakingPoolLogicV2 {
     StakingPoolV2.PoolData storage poolData,
     uint256 rewardPerSecond,
     uint256 roundStartTimestamp,
-    uint32 duration
+    uint256 duration
   ) internal returns (uint256, uint256) {
     poolData.rewardPerSecond = rewardPerSecond;
     poolData.startTimestamp = roundStartTimestamp;
     poolData.endTimestamp = roundStartTimestamp + duration;
     poolData.lastUpdateTimestamp = roundStartTimestamp;
     poolData.rewardIndex = 1e18;
-    poolData.duration = duration;
+    poolData.duration = uint32(duration);
     return (poolData.startTimestamp, poolData.endTimestamp);
   }
